+++ src/main/java/com/ditrit/letomodelizerapi/model/library/LibraryTemplateType.java
@@ -1,0 +1,23 @@
+package com.ditrit.letomodelizerapi.model.library;
+
+/**
+ * Enum defining the types of library templates.
+ * - PROJECT: Represents a project template.
+ * - DIAGRAM: Represents a diagram template.
+ * - COMPONENT: Represents a component template.
+ */
+public enum LibraryTemplateType {
+
+    /**
+     * Represents a project template.
+     */
+    PROJECT,
+    /**
+     * Represents a diagram template.
+     */
+    DIAGRAM,
+    /**
+     * Represents a component template.
+     */
+    COMPONENT;
+}