+++ src/main/java/com/ditrit/letomodelizerapi/model/library/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains the model classes related to library in the Ditrit LetoModelizer API.
+ * It includes various entities, enums, and other data structures used to manage and represent library mechanisms.
+ */
+package com.ditrit.letomodelizerapi.model.library;