+++ src/main/resources/db/migration/V1_6_0__new_table_permissions.sql
@@ -1,5 +1,5 @@
 CREATE TYPE entity_type AS ENUM ('ADMIN', 'PROJECT', 'PROJECT_TEMPLATE', 'DIAGRAM', 'DIAGRAM_TEMPLATE', 'COMPONENT', 'COMPONENT_TEMPLATE', 'LIBRARY');
-CREATE TYPE action_type AS ENUM ('ACCESS', 'CREATE', 'DELETE');
+CREATE TYPE action_type AS ENUM ('ACCESS', 'CREATE', 'DELETE', 'UPDATE');
 
 CREATE TABLE IF NOT EXISTS permissions (
     per_id      SERIAL PRIMARY KEY,
@@ -16,21 +16,24 @@ COMMENT ON TYPE   action_type             IS 'An ENUM type representing the diff
 
 COMMENT ON TABLE  permissions             IS 'This table stores permission settings for various entities within the system.';
 COMMENT ON COLUMN permissions.per_id      IS 'Primary key, serial.';
-COMMENT ON COLUMN permissions.entity      IS 'Specifies the type of entity for which the permission is being set, based on the entity_type ENUM. ';
+COMMENT ON COLUMN permissions.entity      IS 'Specifies the type of entity for which the permission is being set, based on the entity_type ENUM.';
 COMMENT ON COLUMN permissions.action      IS 'Defines the action that is permissible for the entity, based on the action_type ENUM.';
 COMMENT ON COLUMN permissions.lib_id      IS 'References the lib_id in the libraries table. This optional field links the permission to a specific library, allowing for permissions to be set at a library-specific level.';
 COMMENT ON COLUMN permissions.insert_date IS 'Creation date of this row.';
 COMMENT ON COLUMN permissions.update_date IS 'Last update date of this row.';
 
 INSERT INTO permissions(per_id, entity, action, lib_id) VALUES
-(1, 'ADMIN', 'ACCESS', NULL),
-(2, 'PROJECT', 'CREATE', NULL),
-(3, 'PROJECT_TEMPLATE', 'CREATE', NULL),
-(4, 'DIAGRAM', 'CREATE', NULL),
-(5, 'DIAGRAM', 'DELETE', NULL),
-(6, 'DIAGRAM_TEMPLATE', 'CREATE', NULL),
-(7, 'COMPONENT', 'CREATE', NULL),
-(8, 'COMPONENT_TEMPLATE', 'CREATE', NULL),
-(9, 'LIBRARY', 'ACCESS', NULL);
+(1,  'ADMIN', 'ACCESS', NULL),
+(2,  'PROJECT', 'CREATE', NULL),
+(3,  'PROJECT_TEMPLATE', 'CREATE', NULL),
+(4,  'DIAGRAM', 'CREATE', NULL),
+(5,  'DIAGRAM', 'DELETE', NULL),
+(6,  'DIAGRAM_TEMPLATE', 'CREATE', NULL),
+(7,  'COMPONENT', 'CREATE', NULL),
+(8,  'COMPONENT_TEMPLATE', 'CREATE', NULL),
+(9,  'LIBRARY', 'ACCESS', NULL),
+(10, 'LIBRARY', 'CREATE', NULL),
+(11, 'LIBRARY', 'DELETE', NULL),
+(12, 'LIBRARY', 'UPDATE', NULL);
 
-ALTER SEQUENCE permissions_per_id_seq RESTART WITH 10;
+ALTER SEQUENCE permissions_per_id_seq RESTART WITH 100;